Rename namelist.py to _namelist.py
