+++ esmvaltool/_main.py
@@ -47,7 +47,7 @@ from multiprocessing import cpu_count
 
 from . import __version__
 from ._config import configure_logging, read_config_user_file
-from .namelist import read_namelist_file
+from ._namelist import read_namelist_file
 from .task import resource_usage_logger
 
 # set up logging
